Pnpm (#415)
* Switch to pnpm for package management

- Updated UI application addition from `AddNpmApp` to `AddPnpmApp` for improved dependency handling.
- Added `preinstall` script in `package.json` to enforce `pnpm` usage.
- Ensured consistency in dependency management practices across the project with updates to `pnpm-lock.yaml`.

* Update main.yml

* Update main.yml

* Update main.yml

* Update main.yml

* Update pnpm caching and package manager configuration

- Added caching for pnpm store in main.yml to improve installation speed.
- Introduced a step to enable pnpm with corepack.
- Removed the previous cache step, restructuring caching strategy.
- Updated package.json to specify pnpm as the package manager (v10.9.0) with SHA-512 checksum.

* Remove --if-present flag from pnpm commands

Updated `main.yml` to remove the `--if-present` flag from the pnpm commands for generating OpenAPI documentation, linting, and building. This change means these commands will now run unconditionally, which may lead to errors if the corresponding scripts are not defined in package.json.

* Remove preinstall script and package-lock.json

This commit removes the `preinstall` script from `package.json`, which enforced the use of `pnpm` as the package manager. It also deletes the `package-lock.json` file, suggesting a shift away from using a lockfile for dependency management.

* Remove preinstall script from package.json as it didn't work

Eliminate the enforcement of pnpm as the package manager
during the pre-installation phase by removing the
preinstall script from the build configuration. It didn't work